Add TimeTrait to consolidate timestamp handling
