+++ Core/src/GrpcTrait.php
@@ -33,6 +33,7 @@ use Google\Protobuf\NullValue;
 trait GrpcTrait
 {
     use ArrayTrait;
+    use TimeTrait;
     use WhitelistTrait;
 
     /**
@@ -105,17 +106,6 @@ trait GrpcTrait
             'transport' => 'grpc',
             'authHttpHandler' => $authHttpHandler
         ];
-    }
-
-    /**
-     * Format a gRPC timestamp to match the format returned by the REST API.
-     *
-     * @param array $timestamp
-     * @return string
-     */
-    private function formatTimestampFromApi(array $timestamp)
-    {
-        return Timestamp::formatArrayAsString($timestamp);
     }
 
     /**
@@ -236,6 +226,24 @@ trait GrpcTrait
     }
 
     /**
+     * Format a gRPC timestamp to match the format returned by the REST API.
+     *
+     * @param array $timestamp
+     * @return string
+     */
+    private function formatTimestampFromApi(array $timestamp)
+    {
+        $timestamp += [
+            'seconds' => 0,
+            'nanos' => 0
+        ];
+
+        $dt = $this->createDateTimeFromSeconds($timestamp['seconds']);
+
+        return $this->formatTimeAsString($dt, $timestamp['nanos']);
+    }
+
+    /**
      * Format a timestamp for the API with nanosecond precision.
      *
      * @param string $value
@@ -243,14 +251,11 @@ trait GrpcTrait
      */
     private function formatTimestampForApi($value)
     {
-        preg_match('/\.(\d{1,9})Z/', $value, $matches);
-        $value = preg_replace('/\.(\d{1,9})Z/', '.000000Z', $value);
-        $dt = \DateTimeImmutable::createFromFormat('Y-m-d\TH:i:s.u\Z', $value, new \DateTimeZone('UTC'));
-        $nanos = (isset($matches[1])) ? str_pad($matches[1], 9, '0') : 0;
+        list ($dt, $nanos) = $this->parseTimeString($value);
 
         return [
-            'seconds' => (int)$dt->format('U'),
-            'nanos' => (int)$nanos
+            'seconds' => (int) $dt->format('U'),
+            'nanos' => (int) $nanos
         ];
     }
 }
